README updates (#37)
📖 Small corrections, updates to the latest API, and grammar tweaks.

Motivation:

Make sure our README is current and easy to follow.

Modifications:

Graphite is a backend for metrics, Grafana is a frontend for visualization+alerting that can be powered by many backends.
We have a tagged release.
Add a link to @MrLotU 's SwiftPrometheus (emoji optional 😄 )
Some (mostly optional) punctuation tweaks
Tried to re-word the "applications choose their backend" section so it does not use the passive voice.
Update to the latest API... do we still want to write this out this here vs. link to the auto-generated API docs? 🤔 I'm leaning toward deleting it, but there's some value in it all being present here so long as we remember to update it.
Result:

Ideally, a better README! 📖