Add documentation and commit hook for SwiftFormat.
Motivation:

We enforce the use of SwiftFormat in our `sanity.sh` script,
but do not mention that in our documentation. We should point
out the usage in our Contributing guidelines, and also a helper
to automatically run it as a git commit.

Modifications:

Added details about SwiftFormat, a commit hook, and also the
.swift-version to make sure SwiftFormat is able to run based on
the base version of Swift we want to support.

Result:

Contributors will be aware of the formatting requirements of the repo.