TestMetrics.FullKey hash goes via Dictionary (#179)
`TestMetrics.FullKey` `hash` goes via intermediate `Dictionary`.

### Motivation:

Recently a change was made to the `TestMetrics.FullKey.hash` method to
make the hashing output stable and remove its dependency on the order of
the key-value pairs.

### Modifications:

This PR retains that property but instead of sorting first inserts the
key-value pairs into a `Dictionary` as is done in the `==` method.

### Result:

This is useful because it more clearly indicates that the type will
comply with the requirement that anything that is equal must produce the
same hash.